requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
sqlparse==0.5.1
tzdata==2024.1
psycopg2-binary
psycopg_binary
psycopg
django-environ